propose from file
This is working except hevm doesn't seem to like taking long strings as
input:

$ libexec/radgov/radgov-propose example
hevm: could not parse abi argument: This is an example proposal for `radgov propose`. The header contains an
arbitrary string description of the proposal followed by a single line
containing `## PROPOSAL ##` followed by the proposal actions in the format
`target value signature calldata`.  The target and calldata must be hex. The
signature is either an abi string or hexdata. The value is either a number or
hexdata. All hexdata must be prefixed with 0x.

For example:

```
0xac0000000000000000000000000000000000ab 0 "f1()"
0xba0000000000000000000000000000000000be 0 "f2(bool)" true
0xca0000000000000000000000000000000000fe 0 "f3(uint256,bytes)" 1 0x1234
0xde0000000000000000000000000000000000ed 0 "f4(uint256[],bool[])" "[1, 2]" "[true, false]"
```

This proposal can be proposed with `radgov propose example`.

Find out more at <https://radicle.xyz> : string
CallStack (from HasCallStack):
  error, called at src/EVM/ABI.hs:512:8 in hevm-0.44.1-57RYt6zhwW0KHeyn7oDeT9:EVM.ABI